Protections against possible problems from numpy 1.17.0 clip()
numpy changed the behavior of clip() in 1.17.0 where it returns
a float when operating on an int array.  This PR provides workarounds
for potential code sections that could be affected.

+++ ginga/ColorDist.py
@@ -35,7 +35,7 @@ class ColorDistBase(object):
     def hash_array(self, idx):
         # NOTE: data could be assumed to be in the range 0..hashsize-1
         # at this point but clip as a precaution
-        idx = idx.clip(0, self.hashsize - 1)
+        idx = idx.clip(0, self.hashsize - 1).astype(np.uint, copy=False)
         arr = self.hash[idx]
         return arr
